Allow Calculator To Input Float

+++ calc/calc.py
@@ -7,7 +7,7 @@ def application(environ, start_response):
 	second_num = d.get('second_num', [''])[0]
 	sum, mul = 0, 0
 	try:
-		first_num, second_num = int(first_num), int(second_num)
+		first_num, second_num = float(first_num), float(second_num)
 		sum = first_num + second_num
 		mul = first_num * second_num
 	except ValueError:
